Refactor(mentor): API 서비스 통합
