+++ lib/api/ArticleService.js
@@ -55,12 +55,36 @@ export async function deleteArticle(id) {
   return res.data;
 }
 
+export async function getArticleCommentList(articleId) {
+  const res = await get(`/articles/${articleId}/comments`);
+  return res.data;
+}
+
+export async function createArticleComment(articleId, comment) {
+  const res = await post(`/articles/${articleId}/comments`, comment);
+  return res.data;
+}
+
+export async function patchArticleComment(articleId, id, comment) {
+  const res = await patch(`/articles/${articleId}/comments/${id}`, comment);
+  return res.data;
+}
+
+export async function deleteArticleComment(articleId, id) {
+  const res = await remove(`/articles/${articleId}/comments/${id}`);
+  return res.data;
+}
+
 const articles = {
   getArticleList,
   getArticle,
   createArticle,
   patchArticle,
-  deleteArticle
+  deleteArticle,
+  getArticleCommentList,
+  createArticleComment,
+  patchArticleComment,
+  deleteArticleComment
 };
 
 export default articles;
